+++ src/main/java/zaftnotameni/creatania/registry/Fluids.java
@@ -38,6 +38,11 @@ public class Fluids {
   public static final RegistryObject<FlowingFluid> CORRUPT_MANA_FLOWING
     = INDEX.register(Constants.CORRUPT_MANA_FLOWING, () -> new ForgeFlowingFluid.Flowing(Fluids.CORRUPT_MANA_FLUID_PROPERTIES));
 
+  public static final RegistryObject<FlowingFluid> BOTANIA_MANA_FLUID
+    = INDEX.register(Constants.BOTANIA_MANA_FLUID, () -> new ForgeFlowingFluid.Source(Fluids.BOTANIA_MANA_FLUID_PROPERTIES));
+  public static final RegistryObject<FlowingFluid> BOTANIA_MANA_FLOWING
+    = INDEX.register(Constants.BOTANIA_MANA_FLOWING, () -> new ForgeFlowingFluid.Flowing(Fluids.BOTANIA_MANA_FLUID_PROPERTIES));
+
   public static final ForgeFlowingFluid.Properties MANA_FLUID_PROPERTIES = new ForgeFlowingFluid.Properties(
     () -> MANA_FLUID.get(),
     () -> MANA_FLOWING.get(),
@@ -52,6 +57,21 @@ public class Fluids {
       .levelDecreasePerBlock(2)
       .block(() -> Fluids.MANA_FLUID_BLOCK.get())
       .bucket(() -> Fluids.MANA_FLUID_BUCKET.get());
+
+  public static final ForgeFlowingFluid.Properties BOTANIA_MANA_FLUID_PROPERTIES = new ForgeFlowingFluid.Properties(
+    () -> BOTANIA_MANA_FLUID.get(),
+    () -> BOTANIA_MANA_FLOWING.get(),
+    FluidAttributes.builder(WATER_STILL_RL, WATER_FLOWING_RL)
+      .density(15)
+      .luminosity(2)
+      .viscosity(5)
+      .sound(SoundEvents.HONEY_DRINK)
+      .overlay(WATER_OVERLAY_RL)
+      .color(0xff44ffff))
+    .slopeFindDistance(2)
+    .levelDecreasePerBlock(2)
+    .block(() -> Fluids.BOTANIA_MANA_FLUID_BLOCK.get())
+    .bucket(() -> Fluids.BOTANIA_MANA_FLUID_BUCKET.get());
 
   public static final ForgeFlowingFluid.Properties CORRUPT_MANA_FLUID_PROPERTIES = new ForgeFlowingFluid.Properties(
     () -> CORRUPT_MANA_FLUID.get(),
@@ -74,11 +94,15 @@ public class Fluids {
   public static BucketItem createManaFluidBucket() {
     return new BucketItem(Fluids.MANA_FLUID, new Item.Properties().tab(CreativeModeTabs.CREATANIA_ITEMS).stacksTo(1));
   }
+  public static BucketItem createBotaniaManaFluidBucket() {
+    return new BucketItem(Fluids.BOTANIA_MANA_FLUID, new Item.Properties().tab(CreativeModeTabs.CREATANIA_ITEMS).stacksTo(1));
+  }
   public static final RegistryObject<Item> MANA_FLUID_BUCKET = Items.INDEX.register(
     Constants.MANA_FLUID_BUCKET_NAME, () -> createManaFluidBucket());
   public static final RegistryObject<Item> CORRUPT_MANA_FLUID_BUCKET = Items.INDEX.register(
     Constants.CORRUPT_MANA_FLUID_BUCKET_NAME, () -> createCorruptManaFluidBucket());
-
+  public static final RegistryObject<Item> BOTANIA_MANA_FLUID_BUCKET = Items.INDEX.register(
+    Constants.BOTANIA_MANA_FLUID_BUCKET_NAME, () -> createBotaniaManaFluidBucket());
   public static LiquidBlock createCorruptManaFluidBlock() {
     return new LiquidBlock(() -> CORRUPT_MANA_FLUID.get(), BlockBehaviour.Properties
       .of(Material.LAVA)
@@ -93,10 +117,21 @@ public class Fluids {
       .noCollission()
       .strength(100f));
   }
+  public static LiquidBlock createBotaniaManaFluidBlock() {
+    return new LiquidBlock(() -> BOTANIA_MANA_FLUID.get(), BlockBehaviour.Properties
+      .of(Material.LAVA)
+      .color(MaterialColor.COLOR_LIGHT_BLUE)
+      .noOcclusion()
+      .lightLevel(x -> 4)
+      .noCollission()
+      .strength(100f));
+  }
   public static final RegistryObject<LiquidBlock> MANA_FLUID_BLOCK = Blocks.INDEX.register(
     Constants.MANA_FLUID_BLOCK_NAME, () -> createManaFluidBlock());
   public static final RegistryObject<LiquidBlock> CORRUPT_MANA_FLUID_BLOCK = Blocks.INDEX.register(
     Constants.CORRUPT_MANA_FLUID_BLOCK_NAME, () -> createCorruptManaFluidBlock());
+  public static final RegistryObject<LiquidBlock> BOTANIA_MANA_FLUID_BLOCK = Blocks.INDEX.register(
+    Constants.BOTANIA_MANA_FLUID_BLOCK_NAME, () -> createBotaniaManaFluidBlock());
 
   public static void register(IEventBus bus) {
     Log.LOGGER.debug("register fluids");